feat: turn storage into modules
